tasKing_api/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
